backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn[standard]
openai
torch==2.5.1+cpu
faiss-cpu
adapters==1.1.1
transformers==4.48.3
numpy
networkx
boto3
huggingface_hub
python-dotenv
sse-starlette
